pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "oil-supply-chain-optimizer"
version = "1.0.0"
description = "Multi-period MILP for crude oil supply chain optimization — Permian Basin to Gulf Coast"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
keywords = ["optimization", "milp", "supply-chain", "oil-gas", "operations-research"]

dependencies = [
    "pyomo>=6.7",
    "numpy>=1.26",
    "pandas>=2.2",
    "scipy>=1.12",
    "plotly>=5.22",
    "streamlit>=1.35",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-timeout>=2.3",
    "ruff>=0.4",
    "mypy>=1.10",
    "ipykernel>=6.29",
    "notebook>=7.2",
]

[project.scripts]
osc-run = "run:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--tb=short -q"
timeout = 120

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "W", "I", "N", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
warn_return_any = false

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "data*", "app*"]
